two-final/requirements.txt
crispy-bootstrap4==2022.1
Django==4.1.7
django-baton==2.6.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
Pillow==9.0.1
